chore(weave): move OTel parse types out of trace_server_interface
